shopping cart refactored
